# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "BKRES API",
    "description": "",
    "version": "1.0.0"
  },
  "host": "localhost:5000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to login in a specific user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/user/register": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to create user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "any"
                },
                "account": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "fname": {
                  "example": "any"
                },
                "lname": {
                  "example": "any"
                },
                "role": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/get_user": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to get user",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/user/get_alluser": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to get all user",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "sortBy": {
                  "example": "any"
                },
                "sortOrder": {
                  "example": "any"
                },
                "filterKey": {
                  "example": "any"
                },
                "filterValue": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/deleteuser/{User_key}": {
      "delete": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to delete user",
        "parameters": [
          {
            "name": "User_key",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/user/updateUser/{User_key}": {
      "put": {
        "tags": [
          "User"
        ],
        "description": "Endpoint to update user",
        "parameters": [
          {
            "name": "User_key",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "account": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}